Add a fix to use `--quiet` properly when dynamic term is not present
Sometime, the output from the package can be redirected to something
other than stdout, something like a file in which case getting the
length of the terminal fails and ends up with a
`Inappropriate ioctl for device` error.

Added a fix to get rid of this error by keeping a default value for
terminal length if `--quiet` is passed.